src/docker/docker-flask-app-1/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
mysql-connector-python==8.0.20
numpy==1.18.5
protobuf==3.12.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
yarl==1.4.2
